app/requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.18.0
ollama>=0.1.6
pydantic>=2.6.0
python-jose>=3.3.0
sqlalchemy>=2.0.0
matplotlib>=3.8.0
numpy>=1.24.0
google-cloud-aiplatform[adk,agent_engines]